OAuth: Move data outside of ui branch in the global state tree
