+++ client/login/controller.js
@@ -11,7 +11,7 @@ import qs from 'qs';
 import WPLogin from './wp-login';
 import MagicLogin from './magic-login';
 import HandleEmailedLinkForm from './magic-login/handle-emailed-link-form';
-import { fetchOAuth2ClientData } from 'state/ui/oauth2-clients/actions';
+import { fetchOAuth2ClientData } from 'state/oauth2-clients/actions';
 import { recordTracksEvent } from 'state/analytics/actions';
 
 const enhanceContextWithLogin = context => {
